pyproject.toml
[tool.poetry]
name = "code-runner"
version = "0.1.0"
description = "A code running Discord bot."
authors = ["Lunarmagpie <Bambolambo0@gmail.com>"]
license = "MPLv2"

[tool.poetry.dependencies]
python = ">=3.8,<=3.10"
hikari = "^2.0.0-alpha.108"
hikari-crescent = "^0.1.0"
python-dotenv = "^0.20.0"
aiopyston = "^1.2.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
nox = "^2022.1.7"
mypy = "^0.961"
flake8 = "^4.0.1"
codespell = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
